Update SNAPSHOT handling in release utilities.
Update release.release_publication to accept a --snapshot flag for
indicating when a published release should be marked as such. Logic has
been left in for determining if a release is a snapshot or not based on
tag prefix in case existing build processes rely on this.

Drop entry points for `maven-snapshot-release` and
`python-snapshot-release`. If you want a snapshot release you should
instead call `python-release --snapshot` or the maven equivalent.

Rename python_snapshot_release module to be inline with other support
modules. Update tests as necessary to accommodate.

Resolve #36